pyproject.toml
[project]
name = "tmobile-bill-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosmtplib>=5.1.0",
    "gradio>=6.7.0",
    "matplotlib>=3.10.8",
    "openai>=2.24.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "plotly>=6.5.2",
    "pypdf>=6.7.4",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.47",
]
